:recycle: api: move session usecases into enrolment folder
